requirements.txt
langchain==0.1.1
ipython
streamlit==1.28.2
openai==1.3.4
chromadb==0.4.15
tiktoken
pypdf
streamlit_chat==0.1.1
llama-index==0.9.16.post1
trulens-eval
langchain_experimental==0.0.49
PyMySQL==1.1.0
langchain_community==0.0.13
PyPDF2==3.0.1
google-cloud-storage==2.14.0